+++ docs/docusaurus.config.js
@@ -1,6 +1,6 @@
 /** @type {import('@docusaurus/types').DocusaurusConfig} */
 
-const VERSION = '3.5.0';
+const VERSION = '3.6.0';
 
 module.exports = {
   title: 'Jest Roblox',